# [BEAM-2732][BEAM-4028] Logging relies on StateSampler for context