--- a/sdks/python/apache_beam/runners/worker/statesampler_fast.pyx
+++ b/sdks/python/apache_beam/runners/worker/statesampler_fast.pyx
@@ -90,8 +90,12 @@ cdef class StateSampler(object):
     self.current_state_index = 0
     self.time_since_transition = 0
     self.state_transition_count = 0
-    unknown_state = ScopedState(
-        self, CounterName('unknown'), self.current_state_index)
+    unknown_state = ScopedState(self,
+                                CounterName('unknown'),
+                                None,
+                                self.current_state_index,
+                                None,
+                                None)
     pythread.PyThread_acquire_lock(self.lock, pythread.WAIT_LOCK)
     self.scoped_states_by_index = [unknown_state]
     pythread.PyThread_release_lock(self.lock)
@@ -153,7 +157,7 @@ cdef class StateSampler(object):
   def current_state(self):
     return self.scoped_states_by_index[self.current_state_index]
 
-  cpdef _scoped_state(self, counter_name, output_counter,
+  cpdef _scoped_state(self, counter_name, name_context, output_counter,
                       metrics_container):
     """Returns a context manager managing transitions for a given state.
     Args:
@@ -168,6 +172,7 @@ cdef class StateSampler(object):
     new_state_index = len(self.scoped_states_by_index)
     scoped_state = ScopedState(self,
                                counter_name,
+                               name_context,
                                new_state_index,
                                output_counter,
                                metrics_container)
@@ -183,10 +188,16 @@ cdef class StateSampler(object):
 cdef class ScopedState(object):
   """Context manager class managing transitions for a given sampler state."""
 
-  def __init__(
-      self, sampler, name, state_index, counter=None, metrics_container=None):
+  def __init__(self,
+               sampler,
+               name,
+               step_name_context,
+               state_index,
+               counter,
+               metrics_container):
     self.sampler = sampler
     self.name = name
+    self.name_context = step_name_context
     self.state_index = state_index
     self.counter = counter
     self._metrics_container = metrics_container