--- a/sdks/python/apache_beam/runners/worker/statesampler_slow.py
+++ b/sdks/python/apache_beam/runners/worker/statesampler_slow.py
@@ -21,15 +21,18 @@ from __future__ import absolute_import
 
 from builtins import object
 
+from apache_beam.runners import common
+from apache_beam.utils import counters
+
 
 class StateSampler(object):
 
   def __init__(self, sampling_period_ms):
-    self._state_stack = [ScopedState(None, self, None)]
+    self._state_stack = [ScopedState(self,
+                                     counters.CounterName('unknown'),
+                                     None)]
     self.state_transition_count = 0
     self.time_since_transition = 0
-    self.started = False
-    self.finished = False
 
   def current_state(self):
     """Returns the current execution state.
@@ -40,9 +43,12 @@ class StateSampler(object):
 
   def _scoped_state(self,
                     counter_name,
+                    name_context,
                     output_counter,
                     metrics_container=None):
-    return ScopedState(self, counter_name, output_counter, metrics_container)
+    assert isinstance(name_context, common.NameContext)
+    return ScopedState(
+        self, counter_name, name_context, output_counter, metrics_container)
 
   def _enter_state(self, state):
     self.state_transition_count += 1
@@ -57,14 +63,16 @@ class StateSampler(object):
     pass
 
   def stop(self):
-    self.finished = True
+    pass
 
 
 class ScopedState(object):
 
-  def __init__(self, sampler, name, counter=None, metrics_container=None):
+  def __init__(self, sampler, name, step_name_context,
+               counter=None, metrics_container=None):
     self.state_sampler = sampler
     self.name = name
+    self.name_context = step_name_context
     self.counter = counter
     self.nsecs = 0
     self.metrics_container = metrics_container