… sdks/python/apache_beam/runners/worker/logger.pxd