The Logging module will no longer implement its own context tracking for step and stage names.

Because it no longer plays part in high-performance operations, I'm removing it and its cython annotations.

Passing PostCommit: https://builds.apache.org/job/beam_PostCommit_Python_Verify/4985/