@pabloem 2018-06-04
r: @aaltay 
This unifies context management in Python, which simplifies further feature work, and also expands the use of NameContext, which should improve the separation of runner and sdk harness.