@pabloem 2018-06-28
r: @charlesccychen 
This unifies context management in Python, so that logging will use the state sampler to retrieve its current state. Also, NameContext is used more widely to improve step name management.