@pabloem 2018-07-03
Also removed the extra path in `statesampler` for `CounterFactory`