@pabloem 2018-07-03
Results of the `map_fn_microbenchmark`, for some floating point reason, it gives per-element cost as zero, but if you observe row-per-row, the cost with changes is slightly lower than the cost on master.

On master:
```
     1 element  0.798268 sec
  1001 elements 1.01165 sec
  2001 elements 1.05419 sec
  3001 elements 1.1398 sec
  4001 elements 1.37623 sec
  5001 elements 1.47872 sec
  6001 elements 1.68769 sec
  7001 elements 1.68809 sec
  8001 elements 1.8503 sec
  9001 elements 2.0606 sec
Fixed cost   0.8104043092640963
Per-element  0.0
R^2          0.9845043457059202
```

With these changes:
```
     1 element  0.796835 sec
  1001 elements 0.952501 sec
  2001 elements 1.01314 sec
  3001 elements 1.17117 sec
  4001 elements 1.31416 sec
  5001 elements 1.3791 sec
  6001 elements 1.54986 sec
  7001 elements 1.68663 sec
  8001 elements 1.71841 sec
  9001 elements 1.93632 sec
Fixed cost   0.8011857992764676
Per-element  0.0
R^2          0.9922532858766798

```