@charlesccychen 2018-07-06
Thanks Pablo!  This LGTM.  Can you rebase?  It looks like this is a great performance improvement too, cutting down processing overhead.