@pabloem 2018-07-09
Tests passing. I'll squash and merge this after lunch.