@pabloem 2018-07-09
Squashed commits and resolved conflicts. Reruning tests.